+++ plugins/main/public/components/common/doc-viewer/doc-viewer.tsx
@@ -4,6 +4,13 @@ import { escapeRegExp } from 'lodash';
 import { i18n } from '@osd/i18n';
 import { FieldIcon } from '../../../../../../src/plugins/opensearch_dashboards_react/public';
 import { EuiFlexGroup, EuiFlexItem } from '@elastic/eui';
+import { FILTER_OPERATOR } from '../data-source';
+import { DocViewTableRowBtnFilterAdd } from './table-row-btn-filter-add';
+import { DocViewTableRowBtnFilterRemove } from './table-row-btn-filter-remove';
+import { DocViewTableRowBtnFilterExists } from './table-row-btn-filter-exists';
+import './doc-viewer.scss';
+import { onFilterCellActions } from '../data-grid/filter-cell-actions';
+import { Filter } from '../../../../../../src/plugins/data/common';
 
 const COLLAPSE_LINE_LENGTH = 350;
 const DOT_PREFIX_RE = /(.).+?\./g;
@@ -15,6 +22,9 @@ export type tDocViewerProps = {
   mapping: any;
   indexPattern: any;
   docJSON: any;
+  filters: Filter[];
+  setFilters: (filters: Filter[]) => void;
+  onFilter?: () => void;
 };
 
 /**
@@ -82,13 +92,39 @@ const DocViewer = (props: tDocViewerProps) => {
   const [fieldRowOpen, setFieldRowOpen] = useState(
     {} as Record<string, boolean>,
   );
-  const { flattened, formatted, mapping, indexPattern, renderFields, docJSON } =
-    props;
+  const {
+    flattened,
+    formatted,
+    mapping,
+    indexPattern,
+    renderFields,
+    docJSON,
+    filters,
+    setFilters,
+    onFilter: onClose,
+  } = props;
+
+  const onFilter = (
+    field: string,
+    operation:
+      | FILTER_OPERATOR.IS
+      | FILTER_OPERATOR.IS_NOT
+      | FILTER_OPERATOR.EXISTS,
+    value?: string | string[],
+  ) => {
+    const _onFilter = onFilterCellActions(
+      indexPattern?.id,
+      filters,
+      setFilters,
+    );
+    _onFilter(field, operation, value);
+    onClose?.();
+  };
 
   return (
     <>
       {flattened && (
-        <table className='table table-condensed osdDocViewerTable'>
+        <table className='table table-condensed wzDocViewerTable'>
           <tbody>
             {Object.keys(flattened)
               .sort()
@@ -99,7 +135,7 @@ const DocViewer = (props: tDocViewerProps) => {
                 const isCollapsed = isCollapsible && !fieldRowOpen[field];
                 const valueClassName = classNames({
                   // eslint-disable-next-line @typescript-eslint/naming-convention
-                  osdDocViewer__value: true,
+                  wzDocViewer__value: true,
                   'truncate-by-height': isCollapsible && isCollapsed,
                 });
                 const isNestedField =
@@ -123,7 +159,7 @@ const DocViewer = (props: tDocViewerProps) => {
 
                 return (
                   <tr key={index} data-test-subj={`tableDocViewRow-${field}`}>
-                    <td className='osdDocViewer__field'>
+                    <td className='wzDocViewer__field'>
                       <EuiFlexGroup
                         alignItems='center'
                         gutterSize='s'
@@ -147,6 +183,35 @@ const DocViewer = (props: tDocViewerProps) => {
                       </EuiFlexGroup>
                     </td>
                     <td>
+                      <div className='wzDocViewer__buttons'>
+                        <DocViewTableRowBtnFilterAdd
+                          disabled={!fieldMapping || !fieldMapping.filterable}
+                          onClick={() =>
+                            onFilter(
+                              field,
+                              FILTER_OPERATOR.IS,
+                              flattened[field],
+                            )
+                          }
+                        />
+                        <DocViewTableRowBtnFilterRemove
+                          disabled={!fieldMapping || !fieldMapping.filterable}
+                          onClick={() =>
+                            onFilter(
+                              field,
+                              FILTER_OPERATOR.IS_NOT,
+                              flattened[field],
+                            )
+                          }
+                        />
+                        <DocViewTableRowBtnFilterExists
+                          disabled={!fieldMapping || !fieldMapping.filterable}
+                          onClick={() =>
+                            onFilter(field, FILTER_OPERATOR.EXISTS)
+                          }
+                          scripted={fieldMapping && fieldMapping.scripted}
+                        />
+                      </div>
                       {renderFields &&
                       renderFields?.find(
                         (field: string) => field?.id === displayName,