+++ plugins/main/public/components/common/wazuh-data-grid/wz-data-grid.tsx
@@ -20,10 +20,10 @@ import {
 } from '../data-grid';
 import { getWazuhCorePlugin } from '../../../kibana-services';
 import {
+  Filter,
   IndexPattern,
   SearchResponse,
 } from '../../../../../../src/plugins/data/public';
-import { useDocViewer } from '../doc-viewer';
 import {
   ErrorHandler,
   ErrorFactory,
@@ -53,6 +53,8 @@ export type tWazuhDataGridProps = {
     pageSize: number;
   }) => void;
   onChangeSorting: (sorting: { columns: any[]; onSort: any }) => void;
+  filters: Filter[];
+  setFilters: (filters: Filter[]) => void;
 };
 
 const WazuhDataGrid = (props: tWazuhDataGridProps) => {
@@ -67,6 +69,8 @@ const WazuhDataGrid = (props: tWazuhDataGridProps) => {
     onChangeSorting,
     query,
     dateRange,
+    filters,
+    setFilters,
   } = props;
   const [inspectedHit, setInspectedHit] = useState<any>(undefined);
   const [isExporting, setIsExporting] = useState<boolean>(false);
@@ -115,11 +119,6 @@ const WazuhDataGrid = (props: tWazuhDataGridProps) => {
     onChangeSorting && onChangeSorting(sorting || []);
   }, [JSON.stringify(sorting)]);
 
-  const docViewerProps = useDocViewer({
-    doc: inspectedHit,
-    indexPattern: indexPattern as IndexPattern,
-  });
-
   const timeField = indexPattern?.timeFieldName
     ? indexPattern.timeFieldName
     : undefined;
@@ -150,6 +149,8 @@ const WazuhDataGrid = (props: tWazuhDataGridProps) => {
     }
   };
 
+  const closeFlyoutHandler = () => setInspectedHit(undefined);
+
   return (
     <>
       {isLoading ? <LoadingSearchbarProgress /> : null}
@@ -178,7 +179,7 @@ const WazuhDataGrid = (props: tWazuhDataGridProps) => {
         </div>
       ) : null}
       {inspectedHit && (
-        <EuiFlyout onClose={() => setInspectedHit(undefined)} size='m'>
+        <EuiFlyout onClose={closeFlyoutHandler} size='m'>
           <EuiFlyoutHeader>
             <EuiTitle>
               <DocDetailsHeader
@@ -197,6 +198,9 @@ const WazuhDataGrid = (props: tWazuhDataGridProps) => {
                     defaultTableColumns,
                     wzDiscoverRenderColumns,
                   )}
+                  filters={filters}
+                  setFilters={setFilters}
+                  onFilter={closeFlyoutHandler}
                 />
               </EuiFlexItem>
             </EuiFlexGroup>