+++ plugins/main/public/components/common/wazuh-discover/wz-flyout-discover.tsx
@@ -270,7 +270,13 @@ const WazuhFlyoutDiscoverComponent = (props: WazuhDiscoverProps) => {
         indexPattern,
       })
     ) : (
-      <DocDetails doc={doc} item={item} indexPattern={indexPattern} />
+      <DocDetails
+        doc={doc}
+        item={item}
+        indexPattern={indexPattern}
+        filters={filters}
+        setFilters={setFilters}
+      />
     );
   };